Fix negative byte length values

+++ CesiumIonRevitAddin/Utils/GLTFExportUtils.cs
@@ -30,7 +30,7 @@ namespace CesiumIonRevitAddin.Utils
             int elementId,
             bool exportNormals)
         {
-            int byteOffset = 0;
+            ulong byteOffset = 0;
 
             var buffer = new GltfBuffer
             {
